remove unused files and view controllers, clean up and organize storyboard, fix key verification crash, fix auth key export crash, change single sig cell color back to grey, automatically segue to create wallet if no wallet exists,
