+++ FullyNoded2/FullyNoded2/Helpers/Tor/TorClient.swift
@@ -349,9 +349,7 @@ class TorClient {
                     let nodesCount = entity!.count
                     
                     for (i, n) in entity!.enumerated() {
-                                                
-                        let node = NodeStruct(dictionary: n)
-                        
+                                                                        
                         cd.retrieveEntity(entityName: .auth) { (authKeys, errorDescription) in
                             
                             if errorDescription == nil {
