+++ FullyNoded2/FullyNoded2/View Controllers/Home Screen/AuthenticateViewController.swift
