+++ FullyNoded2/FullyNoded2/View Controllers/Home Screen/NodeScannerViewController.swift
