+++ FullyNoded2/FullyNoded2/View Controllers/Home Screen/ScannerViewController.swift
