+++ FullyNoded2/FullyNoded2/View Controllers/Wallets/Wallet Tools/QRDisplayerViewController.swift
