+++ FullyNoded2/FullyNoded2/Wallet Logic/Signers/OfflineSignerLegacy.swift
