+++ FullyNoded2/FullyNoded2/Wallet Logic/Transaction Builders/SingleSigBuilder.swift
@@ -47,7 +47,7 @@ class SingleSigBuilder {
         
         func signSegwit(psbt: String) {
             
-            let signer = OfflineSigner()
+            let signer = NativeSegwitOfflineSigner()
             signer.signTransactionOffline(unsignedTx: psbt) { (signedTx) in
 
                 if signedTx != nil {
